pyproject.toml
[project]
name = "cinereserve"
version = "0.1.0"
description = "Projeto Django do CineReserve."
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django>=5.2.12,<6.0.0",
    "djangorestframework>=3.16.1,<4.0.0",
    "python-decouple>=3.8,<4.0.0",
    "psycopg[binary]>=3.3.3,<4.0.0",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
